+++ src/JobSchedulers.jl
@@ -99,15 +99,15 @@ function __init__()
     SINGLE_THREAD_MODE[] = isempty(TIDS)
 
     # initiating THREAD_POOL
-    c = AtomicChannel{Int}(SINGLE_THREAD_MODE[] ? 1 : length(TIDS))
+    c = AtomicChannel{Int,true}(SINGLE_THREAD_MODE[] ? 1 : length(TIDS))
     global THREAD_POOL[] = c
     for i in TIDS
         put!(c, i)
     end
 
     # initiating scheduler action Channel.
-    # global SCHEDULER_ACTION = AtomicChannel{Int}(1)
-    # global SCHEDULER_PROGRESS_ACTION = AtomicChannel{Int}(1)
+    # global SCHEDULER_ACTION = AtomicChannel{Int,true}(1)
+    # global SCHEDULER_PROGRESS_ACTION = AtomicChannel{Int,true}(1)
 
     # initiating JOB ID
     global JOB_ID